move tei_models into source_model
