--- .sbt_metadata/tei_models.json
@@ -1,6 +1,0 @@
-{
-  "id" : "tei_models",
-  "folder" : "tei_adapter/tei_models",
-  "dependencyIds" : [
-  ]
-}
