+++ common/source_model/src/main/scala/weco/catalogue/source_model/tei/TeiStoreRecord.scala
@@ -1,9 +1,7 @@
-package weco.catalogue.tei.models
+package weco.catalogue.source_model.tei
 
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
 
 import java.time.ZonedDateTime
-
-case class TeiStoreRecord(id: String, metadata: TeiMetadata, version: Int)
 
 case class TeiMetadata(deleted: Boolean, s3Location: S3ObjectLocation, timeModified: ZonedDateTime)
