+++ project/Dependencies.scala
@@ -297,10 +297,6 @@ object CatalogueDependencies {
 
   // TEI adapter
 
-  val teiIdModelsDependencies: Seq[ModuleID] = {
-    WellcomeDependencies.storageLibrary ++
-      WellcomeDependencies.jsonLibrary
-  }
   val teiIdExtractorDependencies: Seq[ModuleID] = {
     WellcomeDependencies.messagingTypesafeLibrary ++
       WellcomeDependencies.httpLibrary ++
